chore create getTradeForWallet selector
