+++ src/components/Trade/hooks/useSwapper/useSwapper.tsx
@@ -1,19 +1,10 @@
 import { type Asset } from '@shapeshiftoss/asset-service'
-import type { UtxoBaseAdapter, UtxoChainId } from '@shapeshiftoss/chain-adapters'
 import { SwapperManager } from '@shapeshiftoss/swapper'
 import { useCallback, useEffect, useMemo, useState } from 'react'
 import { useSelector } from 'react-redux'
 import { getSwapperManager } from 'components/Trade/hooks/useSwapper/swapperManager'
-import {
-  isSupportedCosmosSdkSwappingChain,
-  isSupportedNonUtxoSwappingChain,
-  isSupportedUtxoSwappingChain,
-} from 'components/Trade/hooks/useSwapper/typeGuards'
 import { filterAssetsByIds } from 'components/Trade/hooks/useSwapper/utils'
-import { type BuildTradeInputCommonArgs } from 'components/Trade/types'
-import { getChainAdapterManager } from 'context/PluginProvider/chainAdapterSingleton'
 import { useWallet } from 'hooks/useWallet/useWallet'
-import { toBaseUnit } from 'lib/math'
 import { selectAssetIds } from 'state/slices/assetsSlice/selectors'
 import { selectFeatureFlags } from 'state/slices/preferencesSlice/selectors'
 import {
@@ -22,7 +13,7 @@ import {
   selectPortfolioAccountMetadataByAccountId,
 } from 'state/slices/selectors'
 import { useAppSelector } from 'state/store'
-import { selectQuote, selectSlippage } from 'state/zustand/swapperStore/selectors'
+import { selectGetTradeForWallet, selectQuote } from 'state/zustand/swapperStore/selectors'
 import { useSwapperStore } from 'state/zustand/swapperStore/useSwapperStore'
 
 /*
@@ -34,13 +25,10 @@ export const useSwapper = () => {
   const activeSwapper = useSwapperStore(state => state.activeSwapperWithMetadata?.swapper)
   const sellAssetAccountId = useSwapperStore(state => state.sellAssetAccountId)
   const buyAssetAccountId = useSwapperStore(state => state.buyAssetAccountId)
-  const isSendMax = useSwapperStore(state => state.isSendMax)
   const isExactAllowance = useSwapperStore(state => state.isExactAllowance)
-  const slippage = useSwapperStore(selectSlippage)
-  const receiveAddress = useSwapperStore(state => state.receiveAddress)
   const buyAsset = useSwapperStore(state => state.buyAsset)
   const sellAsset = useSwapperStore(state => state.sellAsset)
-  const sellAmountCryptoPrecision = useSwapperStore(state => state.sellAmountCryptoPrecision)
+  const getTradeForWallet = useSwapperStore(selectGetTradeForWallet)
 
   // Selectors
   const flags = useSelector(selectFeatureFlags)
@@ -120,79 +108,16 @@ export const useSwapper = () => {
     return txid
   }, [activeSwapper, isExactAllowance, activeQuote, wallet])
 
-  const getTrade = useCallback(async () => {
-    if (!sellAsset) throw new Error('No sellAsset')
-    if (!activeSwapper) throw new Error('No swapper available')
-    if (!sellAmountCryptoPrecision) throw new Error('Missing sellTradeAsset.amount')
-    if (!sellAsset) throw new Error('Missing sellAsset')
-    if (!buyAsset) throw new Error('Missing buyAsset')
-    if (!wallet) throw new Error('Missing wallet')
-    if (!receiveAddress) throw new Error('Missing receiveAddress')
-    if (!sellAssetAccountId) throw new Error('Missing sellAssetAccountId')
-    if (!sellAccountBip44Params) throw new Error('Missing sellAccountBip44Params')
-    if (!buyAccountBip44Params) throw new Error('Missing buyAccountBip44Params')
-    if (!sellAccountMetadata) throw new Error('Missing sellAccountMetadata')
-
-    const buildTradeCommonArgs: BuildTradeInputCommonArgs = {
-      sellAmountBeforeFeesCryptoBaseUnit: toBaseUnit(
-        sellAmountCryptoPrecision,
-        sellAsset.precision,
-      ),
-      sellAsset,
-      buyAsset,
-      wallet,
-      sendMax: isSendMax,
-      receiveAddress,
-      slippage,
-    }
-    const sellAssetChainId = sellAsset.chainId
-    if (isSupportedCosmosSdkSwappingChain(sellAssetChainId)) {
-      const { accountNumber } = sellAccountBip44Params
-      const { accountNumber: receiveAccountNumber } = buyAccountBip44Params
-      return activeSwapper.buildTrade({
-        ...buildTradeCommonArgs,
-        chainId: sellAssetChainId,
-        accountNumber,
-        receiveAccountNumber,
-      })
-    } else if (isSupportedNonUtxoSwappingChain(sellAssetChainId)) {
-      const { accountNumber } = sellAccountBip44Params
-      return activeSwapper.buildTrade({
-        ...buildTradeCommonArgs,
-        chainId: sellAssetChainId,
-        accountNumber,
-      })
-    } else if (isSupportedUtxoSwappingChain(sellAssetChainId)) {
-      const { accountType, bip44Params } = sellAccountMetadata
-      const { accountNumber } = bip44Params
-      if (!bip44Params) throw new Error('no bip44Params')
-      if (!accountType) throw new Error('no accountType')
-      const sellAssetChainAdapter = getChainAdapterManager().get(
-        sellAssetChainId,
-      ) as unknown as UtxoBaseAdapter<UtxoChainId>
-      const { xpub } = await sellAssetChainAdapter.getPublicKey(wallet, accountNumber, accountType)
-      return activeSwapper.buildTrade({
-        ...buildTradeCommonArgs,
-        chainId: sellAssetChainId,
-        accountNumber,
-        accountType,
-        xpub,
-      })
-    }
-  }, [
-    sellAsset,
-    activeSwapper,
-    sellAmountCryptoPrecision,
-    buyAsset,
-    wallet,
-    receiveAddress,
-    sellAssetAccountId,
-    sellAccountBip44Params,
-    buyAccountBip44Params,
-    sellAccountMetadata,
-    isSendMax,
-    slippage,
-  ])
+  const getTrade = useCallback(
+    async () =>
+      await getTradeForWallet({
+        wallet,
+        sellAccountBip44Params,
+        sellAccountMetadata,
+        buyAccountBip44Params,
+      }),
+    [wallet, getTradeForWallet, sellAccountBip44Params, sellAccountMetadata, buyAccountBip44Params],
+  )
 
   useEffect(() => {
     ;(async () => {
